@@ -27,7 +27,7 @@
     "@nestjs/platform-express": "^7.4.4",
     "@nestjs/swagger": "^4.6.1",
     "@nestjs/terminus": "^7.0.1",
-    "@webauthn/server": "^0.1.3",
+    "@simplewebauthn/server": "^0.10.3",
     "agenda": "^3.1.0",
     "class-transformer": "^0.3.1",
     "class-validator": "^0.12.2",
@@ -41,7 +41,7 @@
     "helmet": "^4.1.1",
     "i18n": "^0.13.2",
     "mongodb": "^3.6.2",
-    "mongoose": "^5.10.9",
+    "mongoose": "^5.10.10",
     "mqtt": "^4.2.1",
     "node-jose": "^2.0.0",
     "oauth2orize": "^1.11.0",
@@ -64,6 +64,7 @@
   "devDependencies": {
     "@nestjs/cli": "^7.5.1",
     "@nestjs/testing": "^7.4.4",
+    "@simplewebauthn/typescript-types": "^0.10.0",
     "@types/agenda": "^2.0.9",
     "@types/connect-mongo": "^3.1.3",
     "@types/express": "^4.17.8",
@@ -73,20 +74,20 @@
     "@types/joi": "^14.3.4",
     "@types/mongodb": "^3.5.31",
     "@types/mongoose": "^5.7.36",
-    "@types/node": "^14.14.0",
+    "@types/node": "^14.14.2",
     "@types/oauth": "^0.9.1",
     "@types/passport-local": "^1.0.33",
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
-    "eslint": "^7.11.0",
+    "eslint": "^7.12.0",
     "eslint-plugin-jsdoc": "^30.7.3",
     "eslint-plugin-prefer-arrow": "^1.2.2",
-    "jest": "^26.6.0",
+    "jest": "^26.6.1",
     "prettier": "^2.1.2",
     "supertest": "^5.0.0",
-    "ts-jest": "^26.4.1",
-    "ts-loader": "^8.0.6",
+    "ts-jest": "^26.4.2",
+    "ts-loader": "^8.0.7",
     "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.0.3"
