@@ -23,7 +23,7 @@
     "@nestjs/swagger": "^4.6.1",
     "@nestjs/terminus": "^7.0.1",
     "@nestjs/typeorm": "^7.1.4",
-    "aws-sdk": "^2.775.0",
+    "aws-sdk": "^2.778.0",
     "class-transformer": "^0.3.1",
     "class-validator": "^0.12.2",
     "cron": "^1.8.2",
@@ -46,18 +46,18 @@
     "@types/express": "^4.17.8",
     "@types/hapi__joi": "^17.1.6",
     "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.0",
+    "@types/node": "^14.14.2",
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
-    "eslint": "^7.11.0",
+    "eslint": "^7.12.0",
     "eslint-plugin-jsdoc": "^30.7.3",
     "eslint-plugin-prefer-arrow": "^1.2.2",
-    "jest": "^26.6.0",
+    "jest": "^26.6.1",
     "prettier": "^2.1.2",
     "supertest": "^5.0.0",
-    "ts-jest": "^26.4.1",
-    "ts-loader": "^8.0.6",
+    "ts-jest": "^26.4.2",
+    "ts-loader": "^8.0.7",
     "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.0.3"
