@@ -49,18 +49,18 @@
     "@types/express": "^4.17.8",
     "@types/hapi__joi": "^17.1.6",
     "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.0",
+    "@types/node": "^14.14.2",
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
-    "eslint": "^7.11.0",
+    "eslint": "^7.12.0",
     "eslint-plugin-jsdoc": "^30.7.3",
     "eslint-plugin-prefer-arrow": "^1.2.2",
-    "jest": "^26.6.0",
+    "jest": "^26.6.1",
     "prettier": "^2.1.2",
     "supertest": "^5.0.0",
-    "ts-jest": "^26.4.1",
-    "ts-loader": "^8.0.6",
+    "ts-jest": "^26.4.2",
+    "ts-loader": "^8.0.7",
     "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.0.3"
