@@ -13,18 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.6",
+    "@angular/animations": "^10.2.0",
     "@angular/cdk": "^10.2.5",
-    "@angular/common": "^10.1.6",
-    "@angular/compiler": "^10.1.6",
-    "@angular/core": "^10.1.6",
+    "@angular/common": "^10.2.0",
+    "@angular/compiler": "^10.2.0",
+    "@angular/core": "^10.2.0",
     "@angular/flex-layout": "^9.0.0-beta.31",
-    "@angular/forms": "^10.1.6",
+    "@angular/forms": "^10.2.0",
     "@angular/http": "^7.2.16",
     "@angular/material": "^10.2.5",
-    "@angular/platform-browser": "^10.1.6",
-    "@angular/platform-browser-dynamic": "^10.1.6",
-    "@angular/router": "^10.1.6",
+    "@angular/platform-browser": "^10.2.0",
+    "@angular/platform-browser-dynamic": "^10.2.0",
+    "@angular/router": "^10.2.0",
     "@thebespokepixel/es-tinycolor": "^2.1.1",
     "angular-oauth2-oidc": "^10.0.3",
     "angular-oauth2-oidc-jwks": "^9.0.0",
@@ -34,20 +34,20 @@
     "rxjs": "^6.6.3",
     "tinycolor2": "^1.4.2",
     "tslib": "^2.0.3",
-    "zone.js": "^0.11.2"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1001.7",
-    "@angular/cli": "^10.1.7",
-    "@angular/compiler-cli": "^10.1.6",
-    "@angular/language-service": "^10.1.6",
-    "@types/jasmine": "^3.5.14",
+    "@angular-devkit/build-angular": "^0.1002.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.0",
+    "@angular/language-service": "^10.2.0",
+    "@types/jasmine": "^3.6.0",
     "@types/jasminewd2": "^2.0.8",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
     "codelyzer": "^6.0.1",
     "concurrently": "^5.3.0",
-    "eslint": "^7.11.0",
+    "eslint": "^7.12.0",
     "eslint-plugin-jsdoc": "^30.7.3",
     "eslint-plugin-prefer-arrow": "^1.2.2",
     "jasmine-core": "^3.6.0",
