@@ -13,39 +13,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.3",
-    "@angular/cdk": "^10.2.2",
-    "@angular/common": "^10.1.3",
-    "@angular/compiler": "^10.1.3",
-    "@angular/core": "^10.1.3",
+    "@angular/animations": "^10.2.0",
+    "@angular/cdk": "^10.2.5",
+    "@angular/common": "^10.2.0",
+    "@angular/compiler": "^10.2.0",
+    "@angular/core": "^10.2.0",
     "@angular/flex-layout": "^9.0.0-beta.31",
-    "@angular/forms": "^10.1.6",
+    "@angular/forms": "^10.2.0",
     "@angular/http": "^7.2.16",
-    "@angular/localize": "^10.1.6",
+    "@angular/localize": "^10.2.0",
     "@angular/material": "^10.2.5",
-    "@angular/platform-browser": "^10.1.6",
-    "@angular/platform-browser-dynamic": "^10.1.6",
-    "@angular/platform-server": "^10.1.6",
-    "@angular/router": "^10.1.6",
-    "@webauthn/client": "^0.1.3",
+    "@angular/platform-browser": "^10.2.0",
+    "@angular/platform-browser-dynamic": "^10.2.0",
+    "@angular/platform-server": "^10.2.0",
+    "@angular/router": "^10.2.0",
+    "@github/webauthn-json": "^0.5.7",
     "core-js": "^3.6.5",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.6.3",
     "tslib": "^2.0.3",
-    "zone.js": "^0.11.2"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1001.7",
-    "@angular/cli": "^10.1.7",
-    "@angular/compiler-cli": "^10.1.6",
-    "@angular/language-service": "^10.1.6",
-    "@types/jasmine": "^3.5.14",
+    "@angular-devkit/build-angular": "^0.1002.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.0",
+    "@angular/language-service": "^10.2.0",
+    "@types/jasmine": "^3.6.0",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^14.14.0",
+    "@types/node": "^14.14.2",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
     "codelyzer": "^6.0.1",
-    "eslint": "^7.11.0",
+    "eslint": "^7.12.0",
     "eslint-plugin-jsdoc": "^30.7.3",
     "eslint-plugin-prefer-arrow": "^1.2.2",
     "jasmine-core": "^3.6.0",
