@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { switchMap, map } from 'rxjs/operators';
+import { switchMap } from 'rxjs/operators';
 import { from } from 'rxjs';
-import { solveRegistrationChallenge } from '@webauthn/client';
+import { create } from '@github/webauthn-json';
 
 @Injectable({
   providedIn: 'root',
@@ -21,24 +21,14 @@ export class AuthenticationKeysService {
         { headers: registerReqHeaders },
       )
       .pipe(
-        map(challenge => {
-          const { excludeCredentials } = challenge;
-          return {
-            ...challenge,
-            excludeCredentials: excludeCredentials.map(cred => ({
-              ...cred,
-              id: this.base64ToArrayBuffer(cred.id),
-            })),
-          };
-        }),
         switchMap(challenge => {
-          return from(solveRegistrationChallenge(challenge));
+          return from(create({ publicKey: challenge }));
         }),
         switchMap(credentials => {
-          const challengeReqHeaders = new HttpHeaders().set(
-            'Content-Type',
-            'application/json',
-          );
+          const challengeReqHeaders = new HttpHeaders()
+            .append('Content-Type', 'application/json')
+            .append('Authorization', 'Bearer ' + accessToken);
+
           return this.http.post<{ registered: string }>(
             '/webauthn/v1/register',
             credentials,
@@ -72,16 +62,6 @@ export class AuthenticationKeysService {
       {},
       { headers },
     );
-  }
-
-  base64ToArrayBuffer(base64string: string) {
-    const binaryString = atob(base64string);
-    const len = binaryString.length;
-    const bytes = new Uint8Array(len);
-    for (let i = 0; i < len; i++) {
-      bytes[i] = binaryString.charCodeAt(i);
-    }
-    return bytes.buffer;
   }
 
   getAuthenticators(userUuid: string, accessToken: string) {
