@@ -11,5 +11,11 @@ export interface UserAuthenticator extends Document {
 }
 
 export enum Fmt {
-  packed = 'packed',
+  PACKED = 'packed',
+  FIDO_U2F = 'fido-u2f',
+  ANDROID_SAFETYNET = 'android-safetynet',
+  ANDROID_KEY = 'android-key',
+  TPM = 'tpm',
+  APPLE = 'apple',
+  NONE = 'none',
 }
